Change the options arrays for dispatchers to shapes
It's currently not possible to specify the cacheDisabled option, as it
is a bool. To fix, actually define the acceptable options and their
types.

HHVM allows nullable shape fields to not be specified.

This is BC breaking, as it must now be passed in as a shape instead of
an array.

A non-BC-breaking alternative would be to change the array to
`array<string, mixed>`.